Refactoring to clea up padding
