+++ src/aqsc/MHD_parsed/eval_II_center.py
@@ -1,7 +1,13 @@
 # Evaluating the m=0 component of II, which serevs as the m=0 component
 # of the looped equation at odd orders.
 from math import floor, ceil
-from aqsc.math_utilities import *
+from aqsc.math_utilities import (
+    diff,
+    is_integer,
+    is_seq,
+    py_sum,
+    py_sum_parallel,
+)
 # from jax import jit
 # from functools import partial
 # @partial(jit, static_argnums=(0,))
