+++ src/aqsc/MHD_parsed/eval_loop.py
@@ -5,7 +5,13 @@
 # Must be evaluated with Z_coef_cp[n] = 0, p_perp_coef_cp[n] = 0
 # B_psi_coef_cp[n-2] = 0, B_denom_coef_c[n] = 0 and B_theta_coef_cp[n] = 0
 from math import floor, ceil
-from aqsc.math_utilities import *
+from aqsc.math_utilities import (
+    diff,
+    is_integer,
+    is_seq,
+    py_sum,
+    py_sum_parallel,
+)
 # from jax import jit
 # from functools import partial
 # @partial(jit, static_argnums=(0,))
