+++ src/aqsc/MHD_parsed/eval_p_perp_n.py
@@ -7,7 +7,13 @@
 # coeff = -2diff(Delta_coef_cp[0],'phi',1)/(n*B_alpha_coef[0]*B_denom_coef_c[0])
 # coeff_dphi = -2(Delta_coef_cp[0]-1)/(n*B_alpha_coef[0]*B_denom_coef_c[0])
 from math import floor, ceil
-from aqsc.math_utilities import *
+from aqsc.math_utilities import (
+    diff,
+    is_integer,
+    is_seq,
+    py_sum,
+    py_sum_parallel,
+)
 # from jax import jit
 # from functools import partial
 # @partial(jit, static_argnums=(0,))
