+++ src/aqsc/MHD_parsed/validate.py
@@ -1,6 +1,12 @@
 # This script evaluates all governing equations to order n.
 from math import floor, ceil
-from aqsc.math_utilities import *
+from aqsc.math_utilities import (
+    diff,
+    is_integer,
+    is_seq,
+    py_sum,
+    py_sum_parallel,
+)
 # from jax import jit
 # from functools import partial
 # @partial(jit, static_argnums=(0,))
