+++ src/aqsc/chiphifunc.py
@@ -17,10 +17,6 @@ if double_precision:
     from jax import config as jax_config
     jax_config.update("jax_enable_x64", True)
 
-# Maximum allowed asymptotic series order for y'+py=f
-# This feature is depreciated and no longer included in py.
-asymptotic_order = 6
-
 ''' I. Representing functions of chi and phi (ChiPhiFunc subclasses) '''
 # Represents a function of chi and phi.
 # Manages an complex128[m, n] 2d array called content.
@@ -45,151 +41,6 @@ asymptotic_order = 6
 # 2. Grid rep favors multiplications, while Fourier rep favors diffs and integrals.
 # Multiplications are simply more common.
 # The number of phi modes will be tracked and cleaned up with a low-pass filter.
-
-''' I.0 JIT methods used in the grid implementation '''
-def dchi_op(n_dim:int):
-    '''
-    Generate chi differential operator diff_matrix. diff_matrix@f.content = dchi(f). 
-
-    Input: -----
-
-    n_dim: length of Chi series.
-
-    Output: -----
-
-    2d diagonal matrix with elements [-m, -m+2, ... m].
-    '''
-    ind_chi = n_dim-1
-    mode_chi = jnp.linspace(-ind_chi, ind_chi, n_dim)
-    return jnp.diag(1j*mode_chi)
-
-def trig_to_exp_op(n_dim:int):
-    ''' Converts a ChiPhiFunc from trig to exp fourier series. '''
-    ones = jnp.ones(n_dim//2)
-    if n_dim%2==0:
-        arr_diag = jnp.concatenate([0.5j*ones, 0.5*ones])
-        arr_anti_diag = jnp.concatenate([-0.5j*ones, 0.5*ones])
-    if n_dim%2==1:
-        arr_diag = jnp.concatenate([0.5j*ones, jnp.array([0.5]), 0.5*ones])
-        arr_anti_diag = jnp.concatenate([-0.5j*ones, jnp.array([0.5]), 0.5*ones])
-    return(jnp.diag(arr_diag)+jnp.flipud(jnp.diag(arr_anti_diag)))
-
-def exp_to_trig_op(n_dim:int):
-    ''' Converts a ChiPhiFunc from exp to trig fourier series. '''
-    ones = jnp.ones(n_dim//2)
-    # sin:  +0.5je^-ix-0.5je^ix
-    # cos:  +0.5e^-ix+0.5e^ix
-    if n_dim%2==0:
-        arr_diag = jnp.concatenate([-1j*ones, ones])
-        arr_anti_diag = jnp.concatenate([ones, 1j*ones])
-    if n_dim%2==1:
-        arr_diag = jnp.concatenate([-1j*ones, jnp.array([0.5]), ones])
-        arr_anti_diag = jnp.concatenate([ones, jnp.array([0.5]), 1j*ones])
-    return(jnp.diag(arr_diag)+jnp.flipud(jnp.diag(arr_anti_diag)))
-
-def centered_resize_content(content:jnp.ndarray, target_chi:int):
-    '''
-    Symmetrically pad or trim a content array (2d, axis 0 = chi) to have
-    exactly target_chi rows, keeping the chi mode at row content.shape[0]//2
-    (mode 0 for even-order/odd-row-count content, the +1/-1 straddle for
-    odd-order/even-row-count content) at a fixed, shared position regardless
-    of how many rows are added or removed.
-
-    This is representation-agnostic pure array arithmetic: it's the same
-    formula ChiPhiFunc.__add__ uses internally to reconcile mismatched
-    operand widths (dynamic target, derived from max(a.shape, b.shape) at
-    each call) and that ChiPhiFunc.cap_m() uses to trim (dynamic source,
-    target supplied by the caller). ChiPhiFuncPadded reuses this exact
-    formula with a *static* target_chi (chosen once per container, not
-    recomputed from operand shapes on every call), which is what makes its
-    __add__/__mul__/cap_m branch-free. See chiphifunc_padded.py.
-
-    target_chi must share parity with content.shape[0] (both even-order-
-    family or both odd-order-family widths) — this holds for every call site
-    in this codebase, since chi width parity always matches the underlying
-    order's parity, and this function does not check it (matching the
-    permissiveness of the ChiPhiFunc.__add__ code this was extracted from;
-    callers here are trusted, not adversarial).
-
-    Input: -----
-
-    content: a 2d array, chi (axis 0) by phi (axis 1).
-
-    target_chi: desired number of chi rows in the output. Can be larger
-    (pad) or smaller (trim) than content.shape[0], or equal (no-op).
-
-    Output: -----
-
-    A 2d array with target_chi rows.
-    '''
-    len_chi = content.shape[0]
-    diff = target_chi - len_chi
-    if diff == 0:
-        return(content)
-    if diff > 0:
-        pad_before = diff // 2
-        pad_after = diff - pad_before
-        return(jnp.pad(content, ((pad_before, pad_after), (0, 0))))
-    else:
-        clip_before = (-diff) // 2
-        return(content[clip_before : clip_before + target_chi])
-
-def wrap_grid_content_jit(content:jnp.ndarray):
-    '''
-    Used for wrapping grid content. Defined outside the ChiPhiFunc class so that
-    it can be used in @njit compiled methods.
-    Input: -----
-
-    (m, k) 2d array.
-
-    Output: -----
-
-    (m, k+1) 2d array. output[:, -1] = output[:, 0]
-    '''
-    first_col = content[:,0]
-    return(jnp.concatenate((content, first_col[:,None]), axis=1))
-
-def max_log10(input):
-    '''
-    Calculates the max amplitude's order of magnitude
-    Input: -----
-
-    An ndarray.
-
-    Output: -----
-
-    The log10 of the maximum element in input.
-    '''
-    return(jnp.log10(jnp.max(jnp.abs(input)))) # not nfp-sensitive
-
-# Contains static argument.
-# # @partial(jit, static_argnums=(0,))
-def jit_fftfreq_int(int_in:int):
-    '''
-    Shorthand for jnp.fft.fftfreq(n)*n rounded to the nearest int.
-    The input should be static.
-
-    Input: -----
-
-    An integer equal to the length of an array (and its FFT)
-
-    Output: -----
-
-    np.fft.fft_freq(input)*input
-    '''
-    out = jnp.arange(int_in)
-    return(jnp.where(out>(int_in-1)//2,out-int_in,out))
-
-def get_l2_shared(content):
-    power_by_phi = jnp.sum(jnp.abs(content)**2, axis=0)
-    integral_sq = (2*jnp.pi)**2 * jnp.mean(power_by_phi)
-    return(jnp.sqrt(jnp.real(integral_sq)))
-
-'''
-Convolves 2 2d arrays along axis=0.
-The inputs must have with equal lengths along axis=1
-'''
-batch_convolve = vmap(jnp.convolve, in_axes=1, out_axes=1)
 
 def phi_avg(in_quant):
     '''
@@ -1022,19 +873,6 @@ class ChiPhiFunc:
         '''
         return self.pad_chi(m+1)
 
-    def pad_chi_static(self, target_chi:int):
-        '''
-        Like pad_chi, but also handles target_chi < self.content.shape[0]
-        (symmetric trim instead of pad), and never falls back to __add__'s
-        dynamic (runtime-shape-derived) padding math. A thin wrapper around
-        the module-level centered_resize_content(), which is representation-
-        agnostic and shared with ChiPhiFuncPadded (see chiphifunc_padded.py).
-        target_chi must share parity with self.content.shape[0].
-        '''
-        if self.is_special():
-            return(self)
-        return(ChiPhiFunc(centered_resize_content(self.content, target_chi), self.nfp))
-
     def pad_chi(self, target_chi:int):
         '''
         Pads a ChiPhiFunc to be have a given total number of mode components.
@@ -1203,731 +1041,7 @@ tree_util.register_pytree_node(ChiPhiFunc,
                                ChiPhiFunc._tree_flatten,
                                ChiPhiFunc._tree_unflatten)
 
-''' I.2 Utilities '''
-# Used only in pseudosspectral method.
-roll_axis_01 = lambda a, shift: jnp.roll(jnp.roll(a, shift, axis=0), shift, axis=1)
-batch_roll_axis_01 = vmap(roll_axis_01, in_axes=0, out_axes=0)
-# @lru_cache(maxsize=10)
-# @partial(jit, static_argnums=(0,))
-def dphi_op_pseudospectral(n:int):
-    """
-    Return the spectral differentiation matrix for n grid points
-    on the periodic domain [xmax, xmax). This routine is a JAX port of
-    qsc.spectral_diff_matrix in
-    https://github.com/landreman/pyQSC/,
-    and scipy.linalg.toeplitz in
-    https://github.com/scipy/scipy.
-    """
-    h = 2 * jnp.pi / n
-    kk = jnp.arange(1, n)
-    n1 = n//2-(n+1)%2
-    n2 = n//2
-    if n%2 == 0:
-        topc = 1 / jnp.tan(jnp.arange(1, n2 + 1) * h / 2)
-        temp = jnp.concatenate((topc, -jnp.flip(topc[0:n1])))
-    else:
-        topc = 1 / jnp.sin(jnp.arange(1, n2 + 1) * h / 2)
-        temp = jnp.concatenate((topc, jnp.flip(topc[0:n1])))
 
-    # Calculating the Toeplitz matrix
-    col1 = jnp.array([jnp.concatenate((jnp.array([0]), 0.5 * ((-1) ** kk) * temp))]).T
-    col1 = jnp.concatenate([col1, jnp.zeros((n, n-1))], axis=1)
-    row1 = -col1.T
-    # Creating the first row and column
-    raw = col1 + row1
-    # Creating a new axis and shifting the rows and columns
-    toeplitz = jnp.repeat(raw[None,:,:], n, axis=0)
-    shifts = jnp.arange(n)
-    # Rolling the first row and first col
-    toeplitz = batch_roll_axis_01(toeplitz, shifts)
-    # Masking the elements that are looped back
-    masks = jnp.repeat(
-        (shifts[:, None] + shifts[None, :])[None, :, :],
-        n, axis=0
-    )
-    masks = jnp.sign(jnp.where((masks+1)//2<=shifts[:,None,None], 0, masks))
-    # Applying the mask
-    toeplitz = toeplitz*masks
-    # Summing along and collapse the new axis.
-    toeplitz = jnp.sum(toeplitz, axis=0)
-    return(toeplitz)
-
-''' II. Deconvolution ("dividing" chi-dependent terms) '''
-# @partial(jit, static_argnums=(2, 3, ))
-def get_O_O_einv_from_A_B(chiphifunc_A:ChiPhiFunc, chiphifunc_B:ChiPhiFunc, rank_rhs:int, Y1c_mode:bool):
-    '''
-    Get O, O_einv and vector_free_coef that solves the eqaution system
-    O Yn = (A + B dchi) Yn = RHS <=>
-      Yn = O_einv@RHS - (Yn0 or Yn1p) * vec_free_coef
-
-    This is an under-determined problem. Here, Yn is an (n+1)-dim vector.
-    A and B are 2-d vectors. O is a known (n+2, n+1) convolution/differential
-    matrix with A and B as kernels. The RHS (not provided here) is a (n+2) vector
-    with (n) linearly-independent, phi-dependent components.
-
-    Or in code format,
-    Yn_content = jnp.einsum('ijk,jk->ik',O_einv,chiphifunc_rhs_content)\
-        + vec_free * vector_free_coef
-
-
-    Inputs: -----
-
-    A, B: ChiPhiFunc coefficients
-
-    rank_rhs: The number of rows in the RHS.
-
-    Outputs: -----
-
-    O_matrices,
-
-    O_einv,
-
-    vector_free_coef: A (n+1*len_phi)-component array
-
-    Y_nfp: the nfp of Y.
-    '''
-    i_free = (rank_rhs+1)//2 # We'll always use Yn0 or Yn1p as the free var.
-
-    # Mismatched length
-    if chiphifunc_A.content.shape[1] != chiphifunc_B.content.shape[1]\
-    and (chiphifunc_A.content.shape[1]!=1 and chiphifunc_B.content.shape[1]!=1):
-        return(jnp.nan, jnp.nan, jnp.nan, jnp.nan)
-
-    stretch_phi = jnp.zeros((1, max(chiphifunc_A.content.shape[1], chiphifunc_B.content.shape[1])))
-    chiphifunc_A_content = chiphifunc_A.content+stretch_phi
-    chiphifunc_B_content = chiphifunc_B.content+stretch_phi
-
-    # generate the LHS operator O_matrices = (va conv + vb conv dchi)
-    O_matrices = 0
-    A_conv_matrices = conv_tensor(chiphifunc_A_content, rank_rhs+1)
-    O_matrices = O_matrices + A_conv_matrices
-
-    dchi_matrix = dchi_op(rank_rhs+1)
-    B_conv_matrices = conv_tensor(chiphifunc_B_content, rank_rhs+1)
-    O_matrices = O_matrices + jnp.einsum('ijk,jl->ilk',B_conv_matrices,dchi_matrix)
-
-    if Y1c_mode and rank_rhs%2==1:
-        # This O, when inverted, yields an O_einv acting on the trig
-        # representation of RHS and vec_free_coef:
-        # Yn_trig = O_einv@RHS_trig - (Yn0 or Yn1c) * vec_free_coef
-        # O_matrices = O_matrices@trig_to_exp_op(O_matrices.shape[1])
-        O_matrices = jnp.einsum('ijp,jm->imp', O_matrices, trig_to_exp_op(O_matrices.shape[1]))
-        O_matrices = jnp.einsum('ij,jmp->imp', exp_to_trig_op(O_matrices.shape[0]), O_matrices)
-    # It seems difficult to remove this inverse...
-    O_einv = batch_matrix_inv_excluding_col(O_matrices)
-    O_einv = jnp.concatenate((O_einv[:i_free], jnp.zeros((1,O_einv.shape[1],O_einv.shape[2])), O_einv[i_free:]))
-    O_free_col = O_matrices[:,i_free,:]
-    vector_free_coef = jnp.einsum('ijk,jk->ik',O_einv, O_free_col)#A_einv@A_free_col
-    vector_free_coef = vector_free_coef.at[i_free].set(-jnp.ones((vector_free_coef.shape[1])))
-
-    if Y1c_mode and rank_rhs%2==1:
-        #             Yn_trig =             O_einv@            RHS_trig - (Yn0 or Yn1c) *             vec_free_coef
-        # trig_to_exp@Yn_trig = trig_to_exp@O_einv@            RHS_trig - (Yn0 or Yn1c) * trig_to_exp@vec_free_coef
-        #             Yn      = trig_to_exp@O_einv@            RHS_trig - (Yn0 or Yn1c) * trig_to_exp@vec_free_coef
-        #             Yn      = trig_to_exp@O_einv@exp_to_trig@RHS      - (Yn0 or Yn1c) * trig_to_exp@vec_free_coef
-        # O_einv = trig_to_exp_op(O_einv.shape[0])@O_einv@exp_to_trig_op(O_einv.shape[1])
-        O_einv = jnp.einsum('ijp,jm->imp', O_einv, exp_to_trig_op(O_einv.shape[1]))
-        O_einv = jnp.einsum('ij,jmp->imp', trig_to_exp_op(O_einv.shape[0]), O_einv)
-        vector_free_coef = trig_to_exp_op(O_einv.shape[0])@vector_free_coef
-
-    return(O_matrices, O_einv, -vector_free_coef)
-
-'''
-III. Grid 1D deconvolution (used for "dividing" a chi-dependent quantity)
-
-This part solves the pointwise product function problem A*va = B*vb woth unknown va.
-by chi mode matching. It treats both pointwise products as matrix
-products of vectors (components are chi mode coeffs, which are rows in ChiPhiFunc.content)
-with convolution matrices A@va = B@vb,
-Where A, B are (m, n+1) (m, n) or (m, n) (m, n) matrices
-   of rank     (n+1)    (n)    or (n)    (n)
-   va, vb are  (n+1)    (n)    or (n)    (n)   -d vectors
-
-These type of equations underlies power-matched recursion relations:
-a(chi, phi) * va(chi, phi) = b(chi, phi) * vb(chi, phi)
-in grid realization.
-where a, b has the same number of chi modes (let's say o), as pointwise product
-with a, b are convolutions, which are (o+(n+1)-1, n+1) or (o+n-1, n) degenerate
-matrices.
-Codes written in this part are specifically for 1D deconvolution used for ChiPhiFunc.
-'''
-
-''' III.2 va has 1 more component than vb '''
-def batch_matrix_inv_excluding_col(in_matrices:jnp.ndarray):
-    '''
-    Invert an (n,n) submatrix of a (n+2,n+1) rectangular matrix by taking the
-    center n rows and excluding the rank_rhs column. "Taking the center n rows"
-    is motivated by the RHS being rank n-1. For solving for Yn. (a (n+1)-dim
-    vector)
-
-    Input: -----
-
-    in_matrices: (n+2, n+1, len_phi) rank_rhs n matrix to invert
-
-    Return: -----
-
-    (m,m,len_phi) matrix A_inv
-    '''
-    # Checking dimensions
-    if in_matrices.ndim != 3:
-        raise ValueError(
-            f"batch_matrix_inv_excluding_col: expected a 3d array, got ndim={in_matrices.ndim}. "
-            "This used to silently return jnp.nan; converted to a loud error since a shape "
-            "mismatch here usually means a ChiPhiFuncPadded operand reached this exact-rank-"
-            "sensitive linear algebra without being cap_m()'d down to its true width first."
-        )
-
-    n_row = in_matrices.shape[0]
-    n_col = in_matrices.shape[1]
-    n_phi = in_matrices.shape[2]
-    rank_rhs = n_col-1
-    ind_col = (rank_rhs+1)//2
-
-    # Checking shape
-    if n_row-1!=n_col:
-        raise ValueError(
-            f"batch_matrix_inv_excluding_col: expected n_row-1==n_col, got n_row={n_row}, "
-            f"n_col={n_col}. This used to silently return jnp.nan; converted to a loud error "
-            "since a shape mismatch here usually means a ChiPhiFuncPadded operand reached this "
-            "exact-rank-sensitive linear algebra without being cap_m()'d down to its true width "
-            "first."
-        )
-
-    n_clip = (n_row-n_col+1)//2 # How much is the transposed array larger than Yn
-    if n_row<=n_col:
-        raise ValueError(
-            f"batch_matrix_inv_excluding_col: expected n_row>n_col, got n_row={n_row}, "
-            f"n_col={n_col}. Input should have more rows than cols."
-        )
-
-    # Remove specfied column (slightly faster than delete)
-    # and remove extra rows (take n_col-1 rows from the center)
-    rows_to_remove = (n_row-(n_col-1))//2
-    # sub = in_matrices[:,jnp.arange(in_matrices.shape[1])!=i_free,:][rows_to_remove:-rows_to_remove, :, :]
-    sub = jnp.delete(in_matrices, ind_col, axis=1)[rows_to_remove:-rows_to_remove, :, :]
-    sub = jnp.moveaxis(sub,2,0)
-    sqinv = jnp.linalg.inv(sub)
-    sqinv = jnp.moveaxis(sqinv,0,2)
-    padded = jnp.pad(sqinv, ((0,0), (rows_to_remove, rows_to_remove), (0,0)))
-    return(padded)
-
-def get_reduced_square_matrices(in_matrices: jnp.ndarray):
-    '''
-    Extract the square (n,n) matrices for solving linear systems,
-    by excluding a column and slicing the middle rows.
-    '''
-    n_row, n_col, n_phi = in_matrices.shape
-    rank_rhs = n_col - 1
-    ind_col = (rank_rhs + 1) // 2
-
-    rows_to_remove = (n_row - (n_col - 1)) // 2
-    sub = jnp.delete(in_matrices, ind_col, axis=1)[rows_to_remove:-rows_to_remove, :, :]
-    # Shape: (n_col - 1, n_col, n_phi)
-    sub = jnp.moveaxis(sub, 2, 0)  # (n_phi, n_col - 1, n_col)
-    return sub  # shape: (n_phi, n, n+1)
-
-''' III.3 Convolution operator generator and ChiPhiFunc.content wrapper '''
-# A jitted vectorized version of jnp.roll
-roll_axis_0 = lambda a, shift: jnp.roll(a, shift, axis=0)
-batch_roll_axis_0 = vmap(roll_axis_0, in_axes=1, out_axes=1)
-# @partial(jit, static_argnums=(1,))
-def conv_tensor(content:jnp.ndarray, n_dim:int):
-    '''
-    Generate a tensor_coef (see looped_solver.py) convolving a ChiPhiFunc
-    to another along axis 0.
-    For multiplication in FFT space during ODE solves.
-    The convolution is done by:
-    x2_conv_y2 = jnp.einsum('ijk,jk->ik',conv_x2, y2.content)
-
-    Input: -----
-
-    content: A content matrix
-
-    n_dim: The convolution matrices act on a content with n_dim
-    chi components.
-
-    Output: -----
-
-    A (content.shape[0]+n_dim-1, n_dim, content.shape[1]), representing
-    a stack of content.shape[1] convolution matrices. The first two
-    axes represent the row and column of a single convolution matrix.
-    '''
-    len_phi = content.shape[1]
-    content_padded = jnp.concatenate((content, jnp.zeros((n_dim-1, len_phi))), axis=0)
-    content_padded = jnp.tile(content_padded[:, None, :], (1, n_dim, 1))
-    shift = jnp.arange(n_dim)
-    return(batch_roll_axis_0(content_padded, shift[None, :]))
-
-roll_fft_last_axis = lambda a, shift: jnp.roll(a, shift, axis=-1)
-batch_roll_fft_last_axis = vmap(roll_fft_last_axis, in_axes=(-2,0), out_axes=2)
-
-def fft_conv_tensor_batch(source:jnp.array):
-    '''
-    Generates a 4D convolution operator in the phi axis for a 3d "tensor coef"
-    (see looped_solver for explanation).
-
-    Input: -----
-
-    An (a, b, len_phi) array, where a and b represents a matrix acting on the chi
-    dependences of a content
-
-    Output: -----
-
-    An (a, b, len_phi, len_phi) array that acts on a content by transposing
-    axis 1 and 2 and then tensordotting. (see explanation for a "tensor operator"
-    in looped_solver)
-    '''
-    len_phi = source.shape[2]
-    arange = jnp.arange(len_phi)
-    # out = jnp.zeros((len_a, len_b, len_phi, len_phi))#, dtype = jnp.complex128)
-    out = jnp.repeat(source[:,:,None,:], len_phi, axis=2)
-    # Where does 0 elem start/end
-    split = (len_phi+1)//2
-    split_b = jnp.roll(jnp.arange(len_phi%2, len_phi+len_phi%2),split)
-    # Turns out rolling by fftfreq and arange(len_phi) is the same.
-    out = batch_roll_fft_last_axis(out, arange)
-    # Masking
-    split_start = jnp.where(split_b<split, split_b, split) # min
-    split_end = jnp.where(split_b<split, split, split_b) # max
-    mask = jnp.where(
-        jnp.logical_and(arange[None,:]>=split_start[:,None], arange[None,:]<split_end[:,None]),
-        0,1
-    )
-    out = out*mask[None,None,:,:]
-    return(jnp.transpose(out, (0,1,3,2))/len_phi)  # not nfp-dependent
-
-''' IV. Solving linear PDE in phi grids '''
-
-''' IV.1 Solving the periodic linear PDE (a + b * dphi + c * dchi) y = f(phi, chi) '''
-def solve_1d_asym(p_eff, f_eff): # not nfp-dependent
-    '''
-    Solves one linear ODE of form y' + p_eff*y = f_eff.
-
-    Inputs: -----
-
-    p_eff: coefficient of y. 1d array.
-
-    f_eff: RHS. 1d array.
-
-    Outputs: -----
-
-    Solution to the equation
-    when p_eff is 0, y is the anti-derivative of f with zero average.
-    '''
-    if jnp.array(p_eff).ndim==0:
-        p_eff = p_eff*jnp.ones_like(f_eff)
-
-    ai = f_eff/p_eff # f/p
-    # Make an 2d array containing asymptotic series terms.
-    # axis=1 is phi dependence and axis=0 is the term number.
-    # Then, add in the 0-th order term.
-    asym_series = jnp.zeros((asymptotic_order+1, len(ai)))+ai[None, :]
-    for i in range(asymptotic_order):
-        # ai is periodic. We use the non-looped value to ensure
-        # that dphi by fft functions correctly
-        ai_new = -(ChiPhiFunc(ai[None, :],1).dphi().content[0])/p_eff
-        ai = ai_new
-        asym_series = asym_series.at[i+1].set(ai_new)
-
-    # Optimal truncation: Find which term has the smallest maximum amplitude.
-    # And set all following terms to 0.
-    max_amp_for_each_term = jnp.max(jnp.abs(asym_series), axis=1)
-    loc_smallest_max_amp = jnp.argmin(max_amp_for_each_term)
-    asym_series = jnp.where(jnp.arange(len(max_amp_for_each_term))[:, None]>loc_smallest_max_amp, 0, asym_series)
-
-    return(jnp.sum(asym_series, axis=0))
-
-# @partial(jit, static_argnums=(2,))
-def solve_1d_fft(p_eff, f_eff, static_max_freq:int=None): # not nfp-dependent
-    '''
-    Solves one linear ODE of form y' + p_eff*y = f_eff.
-    Assumes non-zero p.
-
-    Inputs: -----
-
-    p_eff: coefficient of y. 1d array.
-    f_eff: RHS. 1d array.
-
-    Outputs: -----
-
-    Solution to the equation.
-    The p_eff == 0 case is handled in solve_ODE by a jnp.where
-    statement.
-    '''
-    len_phi = len(f_eff)
-    if jnp.array(p_eff).ndim==0:
-        p_eff = p_eff*jnp.ones_like(f_eff)
-
-    if static_max_freq is None:
-        target_length = len_phi
-    elif static_max_freq<=0:
-        target_length = len_phi
-    else:
-        target_length = min(len_phi, static_max_freq*2)
-    p_fft = fft_filter(jnp.fft.fft(p_eff), target_length, axis=0)
-    f_fft = fft_filter(jnp.fft.fft(f_eff), target_length, axis=0)
-
-    # Both are (len_phi, len_phi) matrices
-    diff_matrix = fft_dphi_op(target_length)
-    conv_matrix = fft_conv_op(p_fft)
-    # inv_dxpp = jnp.linalg.inv(diff_matrix + conv_matrix)
-    # sln_fft = inv_dxpp@f_fft
-    sln_fft = jnp.linalg.solve(diff_matrix + conv_matrix, f_fft)
-    # operator = lx.MatrixLinearOperator(diff_matrix + conv_matrix)
-    # sln_fft = lx.linear_solve(operator, f_fft).value
-    sln = jnp.fft.ifft(fft_pad(sln_fft, len_phi, axis=0), axis=0)
-
-    return(sln)
-
-solve_1d_fft_batch = vmap(solve_1d_fft, in_axes=(0, 0, None), out_axes=0)
-# @partial(jit, static_argnums=(3,))
-def solve_ODE(coeff_arr, coeff_dp_arr, f_arr:jnp.ndarray, static_max_freq:int=None): # not nfp-dependent
-    '''
-    Solves simple linear first order ODE systems in batch:
-    (coeff_phi d/dphi + coeff) y = f. ( y' + p_eff*y = f_eff )
-    (Dchi = +- m * 1j)
-
-    NOTE:
-    Does not work well for p>10 with zeros or resonant p.
-
-    Inputs: -----
-
-    coeff_arr, coeff_dp_arr, f_arr: Components of the equations as 2d matrices.
-    Axis=0 is equation indices and axis=1 is phi dependences. All quantities are
-    assumed periodic
-
-    static_max_freq: Maximum number of Fourier harmonics used.
-
-    Output: -----
-
-    The solution to the equation system as 2d arrays.
-    '''
-    len_phi = f_arr.shape[1]
-
-    if static_max_freq is None:
-        static_max_freq = len_phi//2
-
-    # Rescale the eq into y'+py=f
-    f_eff = f_arr/coeff_dp_arr
-    p_eff = coeff_arr/coeff_dp_arr
-    # Necessary for integration factor, not necessary for spectral method.
-    # f_eff_scaling = jnp.average(jnp.abs(f_eff))
-    # f_eff = f_eff/f_eff_scaling
-
-    if jnp.array(p_eff).ndim==0:
-        p_eff = p_eff+jnp.zeros_like(f_arr)
-
-    # We always assume f is phi-dependent.
-    if p_eff.shape[1] != f_eff.shape[1]:
-        if p_eff.shape[1]==1:
-            p_eff = p_eff*jnp.ones_like(f_arr)
-        else:
-            return(jnp.nan) # Mismatch
-
-    # Uneven component number
-    if p_eff.shape[0]!=f_eff.shape[0]:
-        return(jnp.nan)
-
-    # Asymptotic series is depreciated. It only works well when
-    # minimum amplitude is large. In this case, the FFT method also works well.
-    out_arr = jnp.where(
-        (jnp.all(p_eff == 0, axis=1))[:, None],
-        ChiPhiFunc(f_eff, nfp=1).\
-            integrate_phi_fft(zero_avg=True).content,
-        solve_1d_fft_batch(p_eff, f_eff, static_max_freq)
-    )
-    # return(out_arr*f_eff_scaling)
-    return(out_arr)
-
-
-''' ONLY USED IN solve_1d(). '''
-# @partial(jit, static_argnums=(0,))
-def fft_dphi_op(len_phi:int):
-    '''
-    ONLY USED IN solve_1d().
-
-    Input: -----
-
-    len_phi:
-    Only used in solve_ODE.
-    remove_zero: A list of booleans. When true, replaces the corresponding element
-    in the second output with a matrix of jnp.nan's.
-
-    Output: -----
-
-    1. A dphi operator acting on the fft of
-    a content along axis=1
-    '''
-    fftfreq = jit_fftfreq_int(len_phi)
-    matrix = jnp.identity(len_phi) * 1j * fftfreq
-    return(matrix)
-
-''' ONLY USED IN solve_1d(). '''
-
-def fft_conv_op(source):
-    '''
-    ONLY USED IN solve_1d().
-    A convolution operator acting convolving
-    a fft of len_phi with source.
-
-    Input: -----
-
-    A 1d array
-
-    Output: -----
-
-    A convoution operators acting on the fft of
-    a content along axis=1. Shape is [len_phi, len_phi].
-    '''
-    # We first turn the axis=0 of source into two axes
-    source_eff = source[None,None,:]
-    tensor_eff = fft_conv_tensor_batch(source_eff)
-    # Collapse the first two axes from a diagonal matrix to a 1d array by summing
-    return(tensor_eff[0][0])
-
-# @partial(jit, static_argnums=(4,))
-def solve_ODE_chi(coeff, coeff_dp, coeff_dc, f, static_max_freq: int):
-    '''
-    For solving the periodic linear 1st order ODE
-    (coeff + coeff_dp*dphi + coeff_dc*dchi) y = f(phi, chi)
-    using Fourier method.
-
-    Input: -----
-
-    coeffs are constants or contents. f is a content (see ChiPhiFunc's description).
-    Contant mode and offset are for evaluating the constant component equation
-    dphi y0 = f0.
-
-    Output: -----
-    y is a ChiPhiFunc's content
-    '''
-    # print('f', f)
-    # print('coeff_dp', coeff_dp)
-    # print('coeff_dc', coeff_dc)
-    len_chi = f.shape[0]
-    # Chi harmonics
-    ind_chi = len_chi-1
-    # Multiplies each row with its corresponding mode number.
-    mode_chi = 1j*jnp.linspace(-ind_chi, ind_chi, len_chi, axis=0)[:,None]
-    coeff_eff = (coeff_dc*mode_chi + coeff)
-
-    return(
-        solve_ODE(coeff_eff, coeff_dp, f, \
-            static_max_freq=static_max_freq)
-    )
-
-# @partial(jit, static_argnums=(2,))
-def solve_dphi_iota_dchi(iota, f, static_max_freq: int):
-    '''
-    For solving the periodic linear 1st order ODE
-    (dphi+iota*dchi) y = f(phi, chi)
-    using Fourier method.
-
-    Input: -----
-
-    iota is a constant and f is a ChiPhiFunc.
-
-    Contant mode and offset are for evaluating the constant component equation
-    dphi y0 = f0.
-
-    Output: -----
-
-    y is a ChiPhiFunc's content
-    '''
-    return(
-        solve_ODE_chi(
-            coeff=0,
-            coeff_dp=1,
-            coeff_dc=iota,
-            f=f,
-            static_max_freq=static_max_freq
-        )
-    )
-
-''' V. utilities '''
-
-''' V.1. Low-pass filter for simplifying tensor to invert '''
-# @partial(jit, static_argnums=(1,2,))
-def fft_filter(fft_in:jnp.ndarray, target_length:int, axis:int): # not nfp-dependent
-    '''
-    Shorten an array in FFT representation to leave only target_length elements.
-    (which correspond to a low-pass filter with k<target_length/2*nfp)
-    by removing the highest frequency modes. The resulting array can be IFFT'ed.
-    Target length is the number of harmonics to keep in one field period.
-
-    Inputs: -----
-
-    fft_in: ndarray to filter. Must already be in FFT representation.
-
-    target_length: length of the output array
-
-    axis: axis to filter along
-
-    Output: -----
-
-    Filtered array.
-    '''
-    if target_length>=fft_in.shape[axis] or target_length<0:
-        return(fft_in)
-    # FFT of an array contains mode amplitude in the order given by
-    # fftfreq(length)*length. For example, for length=7,
-    # [ 0.,  1.,  2.,  3., -3., -2., -1.]
-    left = fft_in.take(indices=jnp.arange(0, (target_length+1)//2), axis=axis)
-    right = fft_in.take(indices=jnp.arange(-(target_length//2), 0), axis=axis)
-    return(jnp.concatenate((left, right), axis=axis)*target_length/fft_in.shape[axis])
-
-# @partial(jit, static_argnums=(1,2,))
-def fft_pad(fft_in:jnp.array, target_length:int, axis:int): # not nfp-dependent
-    '''
-    Pad an array in FFT representation to target_length elements.
-    by adding zeroes as highest frequency modes.
-    The resulting array can be IFFT'ed.
-
-    Inputs: -----
-
-    fft_in: ndarray to pad. Must already be in FFT representation.
-
-    target_length: length of the output array
-
-    axis: axis to filter along
-
-    Output: -----
-
-    Padded array.
-    '''
-    if target_length<fft_in.shape[axis]:
-        return(jnp.nan) # Target length smaller than current length
-    elif target_length==fft_in.shape[axis]:
-        return(fft_in)
-    new_shape = list(fft_in.shape)
-    original_length = new_shape[axis]
-    new_shape[axis] = target_length - original_length
-    center_array = jnp.zeros(new_shape)
-    # FFT of an array contains mode amplitude in the order given by
-    # fftfreq(length)*length. For example, for length=7,
-    # [ 0.,  1.,  2.,  3., -3., -2., -1.]
-    left = fft_in.take(indices=jnp.arange(0, (original_length+1)//2), axis=axis)
-    right = fft_in.take(indices=jnp.arange(-(original_length//2), 0), axis=axis)
-    return(jnp.concatenate((left, center_array, right), axis=axis)*target_length/fft_in.shape[axis])
-
-'''
-V.2 Tensor construction for looped_solver.py
-
-Theses methods are for constructing differential/convolution tensors
-'''
-
-# @partial(jit, static_argnums=(1,))
-def to_tensor_fft_op(ChiPhiFunc_in:ChiPhiFunc, len_tensor:int):
-    '''
-    For solving the looped equations. They are only used in looped_solver.py and
-    lambda_coefs_B_psi.py.
-    '''
-    tensor_coef = ChiPhiFunc_in.content[:, None, :]
-    tensor_fft_coef = fft_filter(jnp.fft.fft(tensor_coef, axis = 2), len_tensor, axis=2)
-    tensor_fft_op = fft_conv_tensor_batch(tensor_fft_coef)
-    return(tensor_fft_op)
-
-# @partial(jit, static_argnums=(1,2,3,4,5,6))
-def to_tensor_fft_op_multi_dim(
-    ChiPhiFunc_in:ChiPhiFunc, dphi:int, dchi:int,
-    num_mode:int, cap_axis0:int,
-    len_tensor: int,
-    nfp: int):
-    '''
-    (n, n-2, len_tensor, len_tensor), acting on the FFT of
-    [
-        [B_theta +n-1],
-        ...
-        [B_theta -n+1],
-    ]
-    Generating convolution tensors from B_theta coefficients.
-    These are only needed for n>2 (n_eval>3)
-
-    Inputs: -----
-
-    num_mode: the number of columns of the resulting tensor
-    (corresponds to input row number)
-
-    cap_axis0: the length of axis=0 for the resulting tensor,
-    used to remove outer components that are known to cancel.
-    Must have the same even/oddness and smaller than the row
-    number of the convolution tensor generated from ChiPhiFunc_in
-    and num_mode.
-
-    Output: -----
-
-    A tensor of shape (len_chi+num_mode-1, num_mode, len_tensor, len_tensor)
-    acting on a content by np.tensordot(operator, content, 2).
-    It first takes chi and phi derivatives of specified orders
-    and then multiplies
-    '''
-    if ChiPhiFunc_in.nfp == 0:
-        return(0)
-    # A stack of convolution matrices
-    # shape is (len_chi+num_mode-1, num_mode, len_phi)
-    len_chi = ChiPhiFunc_in.content.shape[0]
-    tensor_coef_nD = conv_tensor(ChiPhiFunc_in.content, num_mode)
-    # Putting in dchi
-    # The outmost component of B_theta is 0.
-    # B_theta coeffs carried by B_psi has 3 components,
-    # and the convolution matrix is n_unknown+2 * n_unknown-1
-    if cap_axis0%2!=tensor_coef_nD.shape[0]%2:
-        return(jnp.full((len_chi+num_mode-1, num_mode, len_tensor, len_tensor),jnp.nan))
-    if cap_axis0>tensor_coef_nD.shape[0]:
-        return(jnp.full((len_chi+num_mode-1, num_mode, len_tensor, len_tensor),jnp.nan))
-    if tensor_coef_nD.shape[0]>cap_axis0:
-        tensor_coef_nD = tensor_coef_nD[
-            (tensor_coef_nD.shape[0]-cap_axis0)//2:
-            (tensor_coef_nD.shape[0]+cap_axis0)//2
-        ]
-    if dchi!=0:
-        dchi_array_temp = (1j*jnp.arange(-num_mode+1,num_mode+1,2)[None, :, None])
-        if dchi>0:
-            tensor_coef_nD = tensor_coef_nD*dchi_array_temp**dchi
-        elif dchi<0:
-            if num_mode%2==0: # chi integrals are only supported when there is no constant componemnt
-                tensor_coef_nD = tensor_coef_nD/dchi_array_temp**(-dchi)
-            else:
-                # This helper method does not support calculating
-                # chi integrals (dchi<0) when the content being acted on
-                # has chi-indep component (num_mode is odd)
-                return(jnp.full((len_chi+num_mode-1, num_mode, len_tensor, len_tensor),jnp.nan))
-    # Applying FFT
-    # A stack of convolution matrices, but now axis=2 is
-    # in frequency space and capped to len_tensor elements.
-    # shape is (len_chi+num_mode-1, num_mode, len_tensor)
-    tensor_fft_coef_B_theta = fft_filter(jnp.fft.fft(tensor_coef_nD, axis = 2), len_tensor, axis=2)
-    # 'Tensor coefficients', dimension is (n_eval-1, n_eval-3, len_phi)
-    # Last 2 dimensions are for convolving phi cells.
-    # shape is (len_chi+num_mode-1, num_mode, len_tensor, len_tensor)
-    tensor_fft_op_B_theta = fft_conv_tensor_batch(tensor_fft_coef_B_theta)
-    # Applying dphi
-    if dphi!=0:
-        if dphi<0:
-            # dphi must be positive
-            return(jnp.full((len_chi+num_mode-1, num_mode, len_tensor, len_tensor),jnp.nan))
-        # dphi matrix
-        fft_freq = jit_fftfreq_int(len_tensor)
-        dphi_array = jnp.ones((len_tensor,len_tensor)) * 1j * fft_freq * nfp
-        tensor_fft_op_B_theta = tensor_fft_op_B_theta*(dphi_array**dphi)
-    return(tensor_fft_op_B_theta)
-
-''' V.3. Others '''
-def linear_least_sq_2d_svd(A, b):
-    '''
-    Solves the linear least square problem minimizing ||Ax-b||
-
-    Inputs: -----
-
-    A, b: (n,m) and (n)
-
-    Output: -----
-
-    x
-    '''
-    u, s, vh = jnp.linalg.svd(A, full_matrices=False)
-    Eps_inv = jnp.diag(1/s)
-    return(vh.conjugate().T@Eps_inv@u.conjugate().T@b)
 
 # Contains static argument.
 # Does not create new instances of the class.
@@ -1972,4 +1086,30 @@ def ChiPhiFuncSpecial(error_code:int):
     # return(ChiPhiFunc(jnp.nan, error_code)) # , is_special=True))
 
 
+# ---------------------------------------------------------------------------
+# Bottom-of-file imports — placed here rather than at the top to break the
+# mutual dependency between chiphifunc and math_utilities:
+#   math_utilities imports ChiPhiFunc (this module) at its top;
+#   chiphifunc methods use the pure helpers below at call time (not
+#   at class-definition time), so fetching them after the class is defined
+#   is safe.  Python returns the partially-initialized math_utilities module
+#   from sys.modules when the circular import resolves, and by that point
+#   all helpers below are already defined there.
+from .math_utilities import (
+    dchi_op,
+    trig_to_exp_op,
+    exp_to_trig_op,
+    dphi_op_pseudospectral,
+    fft_filter,
+    fft_pad,
+    batch_convolve,
+    centered_resize_content,
+    jit_fftfreq_int,
+    wrap_grid_content_jit,
+    get_l2_shared,
+    max_log10,
+    roll_axis_01,
+    batch_roll_axis_01,
+)
+# Existing cycle break: ChiPhiEpsFunc.__add__/__mul__ dispatch back to ChiPhiFunc.
 from .chiphiepsfunc import ChiPhiEpsFunc
