+++ src/aqsc/looped_coefs/B_psi_coefs_in_D3.py
@@ -1,5 +1,12 @@
-from aqsc.math_utilities import *
-from aqsc.chiphifunc import *
+from aqsc.math_utilities import (
+    diff,
+    is_integer,
+    is_seq,
+    py_sum,
+    py_sum_parallel,
+)
+from aqsc.chiphifunc import ChiPhiFuncSpecial
+from aqsc.chiphifunc_solvers import to_tensor_fft_op_multi_dim
 # from jax import jit
 # from functools import partial
 # @partial(jit, static_argnums=(0, 10, 11, 12, 13))
