+++ src/aqsc/looped_coefs/Y_coefs.py
@@ -1,4 +1,10 @@
-from aqsc.math_utilities import *
+from aqsc.math_utilities import (
+    diff,
+    is_integer,
+    is_seq,
+    py_sum,
+    py_sum_parallel,
+)
 
 def eval_y_coefs(n_eval, X_coef_cp, Y_coef_cp,
     Delta_coef_cp, B_alpha_coef, dl_p, tau_p, iota_coef):
