+++ src/aqsc/looped_coefs/Y_coefs_in_D3.py
@@ -1,6 +1,11 @@
 # from math import floor, ceil
-from aqsc.math_utilities import *
-from aqsc.chiphifunc import *
+from aqsc.math_utilities import (
+    diff,
+    is_integer,
+    is_seq,
+    py_sum,
+    py_sum_parallel,
+)
 
 def lambda_coef_Yn1p_in_D3(vector_free_coef, X_coef_cp, Y_coef_cp, iota_coef, dl_p, tau_p, nfp:int):
     coef_Yn1p = (
