+++ src/aqsc/looped_coefs/direct_B_theta_coefs.py
@@ -1,4 +1,10 @@
-from aqsc.math_utilities import *
+from aqsc.math_utilities import (
+    diff,
+    is_integer,
+    is_seq,
+    py_sum,
+    py_sum_parallel,
+)
 
 def eval_B_theta_coefs(p_perp_coef_cp, Delta_coef_cp, B_denom_coef_c, iota_coef):
     coef_B_theta = -(
