+++ src/aqsc/looped_coefs/lambda_B_psi_coef_in_vars.py
@@ -1,4 +1,10 @@
-from aqsc.math_utilities import *
+from aqsc.math_utilities import (
+    diff,
+    is_integer,
+    is_seq,
+    py_sum,
+    py_sum_parallel,
+)
 lambda_B_psi_nm2_in_X_n = lambda n_eval, X_coef_cp, Y_coef_cp, B_theta_coef_cp, B_psi_coef_cp, iota_coef, p_perp_coef_cp, Delta_coef_cp, B_denom_coef_c, B_alpha_coef, kap_p, dl_p, tau_p: \
     -(diff(X_coef_cp[1],True,1)*diff(Y_coef_cp[1],False,1)-iota_coef[0]*X_coef_cp[1]*diff(Y_coef_cp[1],True,2)-X_coef_cp[1]*diff(Y_coef_cp[1],True,1,False,1)-diff(X_coef_cp[1],False,1)*diff(Y_coef_cp[1],True,1)+iota_coef[0]*Y_coef_cp[1]*diff(X_coef_cp[1],True,2)+Y_coef_cp[1]*diff(X_coef_cp[1],True,1,False,1))/(dl_p*kap_p*n_eval-dl_p*kap_p)
 lambda_dphi_B_psi_nm2_in_X_n = lambda n_eval, X_coef_cp, Y_coef_cp, B_theta_coef_cp, B_psi_coef_cp, iota_coef, p_perp_coef_cp, Delta_coef_cp, B_denom_coef_c, B_alpha_coef, kap_p, dl_p, tau_p: \
