+++ src/aqsc/parsed/eval_Znp1.py
@@ -4,7 +4,13 @@
 # iota_coef [..., (n-1)/2 or (n-2)/2]
 # kap_p, dl_p, tau_p
 from math import floor, ceil
-from aqsc.math_utilities import *
+from aqsc.math_utilities import (
+    diff,
+    is_integer,
+    is_seq,
+    py_sum,
+    py_sum_parallel,
+)
 def eval_Znp1_cp(n, X_coef_cp, Y_coef_cp, Z_coef_cp, \
     B_theta_coef_cp, B_psi_coef_cp, \
     B_alpha_coef, \
